combined the two confirm delete modals into a single component
