+++ src/components/index.js
@@ -1,2 +1,3 @@
 export { default as Header } from './header'
 export { default as Footer } from './footer'
+export { default as ResetModal } from './resetModal'
